+++ src/index.ts
@@ -3,10 +3,16 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
+import checkCommitFormat from './checkCommitFormat';
+import getTypeGroup, { GROUPS } from './getTypeGroup';
 import conventionalChangelog from './conventionalChangelog';
 import parserOpts from './parserOpts';
 import recommendedBumpOpts from './recommendedBumpOpts';
 import writerOpts from './writerOpts';
+
+export { GROUPS, checkCommitFormat, getTypeGroup };
+
+export * from './types';
 
 export default Promise.resolve({
   conventionalChangelog,